+++ IL2-SR-Common/Network/ReleaseMetadata.cs
@@ -6,7 +6,7 @@ namespace Ciribob.IL2.SimpleRadio.Standalone.Common.Network
         public const string GithubRepository = "IL2-SimpleRadioStandalone";
         public const string GithubUserAgent = GithubUsername + "_" + GithubRepository;
 
-        public const string Version = "1.0.4.9";
+        public const string Version = "1.0.4.10";
         public const string ReleaseTag = "";
         public const string MinimumProtocolVersion = "1.0.0.0";
     }